+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/zarrs/zarrs/compare/zarrs-v0.23.10...HEAD)
+## [Unreleased](https://github.com/zarrs/zarrs/compare/zarrs-v0.23.11...HEAD)
+
+## [0.23.11](https://github.com/zarrs/zarrs/releases/tag/zarrs-v0.23.11) - 2026-05-13
 
 ### Added
 - Add a `microfloat` feature for expanded subfloat and complex subfloat data type / element support